minor: forge fmt

+++ src/upgradeable/MineFunStorage.sol
@@ -12,34 +12,34 @@ import "./IMineFun.sol";
 abstract contract MineFunStorage is Initializable {
     IUniswapV2Router01 public router;
     address public teamWallet; // Team wallet address
-    mapping(address => uint) public teamFunds; // Tracks ETH allocated for team wallet per token
+    mapping(address => uint256) public teamFunds; // Tracks ETH allocated for team wallet per token
 
     struct MinedToken {
         string name;
         string symbol;
         string metadataCID;
         string tokenImageCID;
-        uint fundingRaised;
-        uint tokensBought;
-        uint bondingDeadline;
+        uint256 fundingRaised;
+        uint256 tokensBought;
+        uint256 bondingDeadline;
         address tokenAddress;
         address creatorAddress;
         bool bonded;
-        uint soloRequiredToMine;
-        mapping(address => uint) contributions; // Tracks user ETH contributions
+        uint256 soloRequiredToMine;
+        mapping(address => uint256) contributions; // Tracks user ETH contributions
     }
 
     address[] public minedTokenAddresses;
     mapping(address => MinedToken) public addressToMinedTokenMapping;
 
     // Constants
-    uint constant MINEDTOKEN_CREATION_FEE = 0.0001 ether;
-    uint constant MINEDTOKEN_FUNDING_GOAL = 1 ether;
-    uint constant PRICE_PER_MINE = 0.0002 ether;
-    uint constant MAX_SUPPLY = 1_000_000_000 ether;
-    uint constant TOKENS_PER_MINE = 50_000 ether;
-    uint constant INIT_SUPPLY = (50 * MAX_SUPPLY) / 100; // 500M tokens
-    uint constant MAX_PER_WALLET = 10_000_000 ether;
+    uint256 constant MINEDTOKEN_CREATION_FEE = 0.0001 ether;
+    uint256 constant MINEDTOKEN_FUNDING_GOAL = 1 ether;
+    uint256 constant PRICE_PER_MINE = 0.0002 ether;
+    uint256 constant MAX_SUPPLY = 1_000_000_000 ether;
+    uint256 constant TOKENS_PER_MINE = 50_000 ether;
+    uint256 constant INIT_SUPPLY = (50 * MAX_SUPPLY) / 100; // 500M tokens
+    uint256 constant MAX_PER_WALLET = 10_000_000 ether;
 
     address public UNISWAP_V2_FACTORY;
     address public UNISWAP_V2_ROUTER;
